# docs/api-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nyamby API",
    "version": "1.0.0",
    "description": "API spec untuk MVP Nyamby — Next.js Route Handlers. Base URL: /api",
    "contact": {
      "name": "Satria Romanda"
    }
  },
  "servers": [
    { "url": "http://localhost:3000/api", "description": "Local Development" },
    { "url": "https://nyamby.vercel.app/api", "description": "Production (Vercel)" }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register user baru",
        "description": "Buat akun baru sebagai talent atau client",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "full_name", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "raka@email.com" },
                  "password": { "type": "string", "minLength": 8, "example": "password123" },
                  "full_name": { "type": "string", "example": "Raka Pratama" },
                  "role": { "type": "string", "enum": ["talent", "client"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registrasi berhasil",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Akun berhasil dibuat",
                  "data": { "id": "uuid", "email": "raka@email.com", "role": "talent" }
                }
              }
            }
          },
          "400": { "description": "Email sudah terdaftar atau validasi gagal" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login berhasil",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "user": { "id": "uuid", "email": "raka@email.com", "role": "talent" },
                    "token": "jwt_token_here"
                  }
                }
              }
            }
          },
          "401": { "description": "Email atau password salah" }
        }
      }
    },
...
    "/ai/match-job": {
      "post": {
        "tags": ["AI"],
        "summary": "🤖 Jalankan AI Job Matching",
        "description": "Dipanggil otomatis dari POST /jobs. Bisa juga dipanggil manual untuk refresh matching. Memanggil GPT-4o dan menyimpan hasilnya ke tabel job_matches.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["job_id"],
                "properties": {
                  "job_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Matching selesai. 12 talenta dievaluasi.",
                  "data": {
                    "job_id": "uuid",
                    "matches_generated": 12,
                    "top_match": { "talent_id": "uuid", "match_score": 87, "recommendation": "highly_recommended" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/skill-gap": {
      "post": {
        "tags": ["AI"],
        "summary": "🤖 Jalankan AI Skill Gap Analysis",
        "description": "Dipanggil otomatis dari POST /talent/onboarding. Bisa dipanggil manual untuk refresh. Memanggil GPT-4o dan menyimpan hasilnya ke tabel skill_gap_analyses.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "talent_profile_id": { "type": "string", "format": "uuid", "description": "Jika tidak diisi, gunakan talenta yang sedang login" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "recommendations": [
                      { "skill_name": "TypeScript", "priority": "high", "reason": "Muncul di 23 dari 30 job aktif", "estimated_impact": "Potensi rate naik 30%" }
                    ],
                    "summary": "Skill kamu sudah solid untuk entry level..."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/escrow/hold": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Client konfirmasi — dana ditahan (mock)",
        "description": "Membuat record escrow_transaction dengan status 'held'. Untuk MVP ini adalah mock — tidak ada integrasi payment gateway.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["job_id", "talent_user_id", "amount"],
                "properties": {
                  "job_id": { "type": "string", "format": "uuid" },
                  "talent_user_id": { "type": "string", "format": "uuid" },
                  "amount": { "type": "number", "example": 2000000 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dana berhasil ditahan. Talenta dapat mulai bekerja.",
                  "data": { "escrow_id": "uuid", "amount": 2000000, "platform_fee": 200000, "status": "held" }
                }
              }
            }
          }
        }
      }
    },
    "/escrow/release": {
      "post": {
        "tags": ["Escrow"],
        "summary": "Client approve — dana dirilis ke talenta (mock)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["job_id"],
                "properties": {
                  "job_id": { "type": "string", "format": "uuid" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dana berhasil dirilis ke talenta. Job selesai!",
                  "data": { "escrow_id": "uuid", "status": "released", "released_at": "2026-05-28T15:00:00Z" }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token dari NextAuth.js. Sertakan di header: Authorization: Bearer <token>"
      }
    }
  }
}